Refactor midonet-api authorisation
This patch removes the typed authorisers in favour of a single
Authorisation class that contains a set of methods to authorise
operations on a given resource.

Removing these per-resource authorisers will allow to easily provide a
different implementation of the Authoriser based on ZOOM in order to
adapt the new backend.

The auth logic is simplified as previously it was requesting data that
it never used (such as the AuthAction) and can potentially be removed
now.  Some logic is also made consistent, such as the behaviour for
authorising non existing resources.  Additionally, ZK access is slightly
improved (e.g., avoid 2 accesses, one in auth, another for the
operation, as the authorisers return the data object after using it).

Change-Id: I1d30b7bb3791bef05ac23c2815f819381d56c8f5
Signed-off-by: Galo Navarro <galo@midokura.com>