Tenant router uplink translation
Updates router gateway translation logic to work with edge routers
rather than the singleton provider router.

This does not address the case where a gateway is created between a
tenant router and an external network with an interface to an edge
router, and then the interface to the edge router is deleted. This
will need to be addressed in a future patch.

Signed-off-by: Brandon Berg <bberg@midokura.com>
Change-Id: Ia500d8aa56d50c62f8c9e57c774125b5a20e3be2