Read port configuration asynchronously
LoadingCache has been updated to allow both synchronous and asynchronous
operation. PortGroupCache has also been changed as it uses
LoadingCache. ReadCache, and the concept of pinning entries has been
removed as it was completely unused.

Synchronous port reading has been left in as it is needed by
ClusterBridgeManager. Otherwise that would also have to be refactored
to be asynchronous and that's a larger change, not justified at this
point.

This change brings the max time waiting on the reactor from ~20s to ~5s
on a setup simultaneously bringing up 20 agents each with 40 ports.

Ref MN-2950

Change-Id: I804c33c39dcd4cfdf34042c2cb9aae1df2495d79
Signed-off-by: Ivan Kelly <ivan@midokura.com>