ArpRequestBrokerTest#testLocallyRepliedArpLoop flake fix
Flake was a result of a very tight race condition. The ArpTable local
map would update, causing get() to return the mac, but the notifiers
wouldn't have been called yet, so the futures would return None.

The fix is to make sure that the notification has happened before
running process().

Ref MI-549

Signed-off-by: Ivan Kelly <ivan@midokura.com>
Change-Id: I7c6a9d64100191246c57512cf360370fc925c531