Check for old host paths when deleting host
the 'command' and 'error' paths under host are no longer
created, so they are ignored when deleting the host. However, if
the deployment went through an upgrade then there is a chance that
the paths exist. In this case, we need to check for the paths and
delete them.

Signed-off-by: Joe Mills <joe@midokura.com>
Fixes MN-3331

Change-Id: I412d37a9ff3af0c9eb2a1f2bf9a2eab2a0bcd0c4