Update patch set 4
Patch Set 4:

This is still WIP.  Some tests are failing.  I will fix them in a bit.  @Galo, the reason why the URI fields must exist as members is because the client side deserialize URIs into these fields.  I updated the patch to differentiate two ways to construct the object:

1. new Foo() -> Jackson deserialization (client)
2. new Foo(baseUri) -> Manual construction (server)

When baseUri is supplied all the URLs are auto-generated in the getters.

Patch-set: 4
Label: Code-Review=0