The MergedMap Trait
This patch introduces the MergedMap trait. Various values associated
to the same key can co-exist in a merged map. These values are called
opinions and each have a specific owner. The merged map relies on a
conflict resolution strategy to determine the winning opinion for each
key.

This patch provides an abstract implementation of the merged map trait
and unit tests for it. This abstract class implements the basic
functionnality to provide an observable and caching for the merged
map. This will simplify the distributed implementation of this trait.

MergedMap and OwnerMap will eventually replace class ReplicatedMap
and are intented to be a more scalable alternative.

Ref: MNA-365 and MNA-366

Change-Id: Ie93c4c7aa26ab2779891fb2adfa5c025678524bf
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>