[WIP] MergedMap and PrivateMap
This patch introduces traits for the merged and private maps.
Each private map has an owner that can put key-value pairs in it.
The merged map merges these private maps into one map according
to some conflict resolution strategy.

Additionally, this patch provides an in-memory implementation of
these traits as well as unit tests.

MergedMap and PrivateMap will eventually replace class ReplicatedMap
and are intented to be a more scalable alternative.

Ref: MNA-365 and MNA-366

Change-Id: Ie93c4c7aa26ab2779891fb2adfa5c025678524bf
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>