The MergedMap
This patch introduces a MergedMap. Various values associated
to the same key can co-exist in a merged map. These values are called
opinions and each have a specific owner. The merged map relies on a
conflict resolution strategy to determine the winning opinion for each
key. This patch provides an implementation of the merged map
and unit tests for it.

The MergedMap will eventually replace class ReplicatedMap
and are intented to be a more scalable alternative.

Ref: MNA-365 and MNA-366

Change-Id: Ie93c4c7aa26ab2779891fb2adfa5c025678524bf
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>