Fixes for virtual topology mapper tests
The patch fixes a bug in the topology mapper unit tests, where in
certain conditions the asserts that were checking whether device
notifications were sent on the VT thread failed. This is because
any subscribers to the VT will receive the initial device, if one
already exists in the VT cache, on the subscribing thread.

While some tests were checking for this exception, the check was
ultimately incompatible with running the tests from multiple
concurrent executors.

To fix the problem, the patch introduces a DeviceObserver, which
is a TestObserver mixed-in with an AwaitableObserver and an
AssertableObserver. The `assert` condition is that notifications
are received either on the VT thread or on the thread where the
observer was created.

Fixes: MNA-369

Change-Id: I825c380ecb12fbeb75b88030df654ad2d9c5f994
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>