Update patch set 1
Patch Set 1: Code-Review+2

Thinking about the case in which there may be more ZK child nodes that we might not be aware of, it would be better to recursively delete from a dir.  That said, I believe such work could be done in a separate work to generally improve the deletion mechanism.  This patch is aligned with how we currently delete many ZK nodes so I will approve.

Patch-set: 1
Label: Code-Review=+2