Add enum validation type to MidoConf
Adds an enum validation type to midoconf. Enum should be defined as a
comma separated list of identifiers inside enum[]. For example,
enum[foo,bar,baz] is an enum with 3 possible values. The identifiers
can only contain letters and numbers. The identifiers are case
_insensitive_, so enum[foo] would validate fine for FOO, FoO fOO etc.

This patch also adds self validation for schemas, so that only valid
enums and validation types can be used.

Change-Id: Ib7559c611d02285cd9354a563a49145a067b7d43
Signed-off-by: Ivan Kelly <ivan@midokura.com>