Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Hi Tomohiko, thanks for the short patch! :-) I'm not an expert on this part of the code, but it looks good to me. The -1 is just because of my doubt about the zoom binding settings, but it would be a +1 if my assumption is false...

Patch-set: 2
Reviewer: Ernest Artiaga <1002750@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1