Address ObservablePathNodeCache flakiness
This patch introduces the TestAwaitableObservable in the
ObservableNodeCacheTest which should help reduce the flakiness in this
test as it's safer than the test observer used earlier.

The patch also makes minor modifications in the tests in order to
prevent race conditions between the creation of test ZK data and subscription.

In local runs, these patches reduce significantly the amount of sporadic
failures.

Additionally, a new assertion is introduced that should help confirm
a hypothesis over the cause of flakiness (not completing with the
NoNodeException as expected).

Ref.: MNA-303
Change-Id: I06f95c6b36c46e770a43b7208c07307f027e87a8
Signed-off-by: Galo Navarro <galo@midokura.com>