API: remove coupling from HTTP exceptions to Agent
This patch cleans a lot of unnecessary code related to http exceptions,
moves them all to the cluster module where they can be shared by all
other modules, and cleans unused code.

As part of this, we remove the dependency from some HTTP exception
classes to Midolman, which was obviously bad coupling.

Change-Id: I915c4d373fefa72bc98b3ef2687fc2af3a49f21d
Signed-off-by: Galo Navarro <galo@midokura.com>