Update patch set 1
Patch Set 1:

(1 comment)

Hi Pino,

> Except: although I can see that you're doing Keystone authentication - I don't see how you're enforcing that ONLY the Admin tenant can access the config API.

...ouch, will add that.

> ConfMinion is already in use in v1.9?

Yes. The legacy REST API is just instantiating, configuring and starting the ConfMinion.

> Does the REST API run the cluster code that runs the minion?

No, just the minion.

> You're simply re-using AuthFilter and AuthContainerRequestFilter for the Config API - great. Is there a good reason to move AuthFilter from midonet-api to midonet-cluster in this patch? Seems orthogonal to authenticating via keystone.

There is, the ConfMinion lives in midonet-cluster and midonet-cluster does not depend on midonet-api. We'd have a circular module dependency if I hadn't moved it.

Patch-set: 1
Label: Code-Review=0