Fixed race condition in async_assert_that
The job to run tcpdump was being dumped into a thread pool and then
a future returned.  That thread pool job then would eventually run
tcpdump and check the output from then on.  However, the main thread
would return control to the main test, which would then run the
send_packet.  It was very possible for the main thread to get to the
send_packet call before the thread pool had executed the tcpdump jobs,
meaning the tcpdump could be run AFTER the send_packet was called
and the packet we were searching for already passed through the system.
This would lead to race conditions and intermittent misses of these
packets, resulting in failing tests.

Plugged up gap for race condition in assert_async by waiting for
a semaphore to be set on the interface object itself which is set only
when tcpdump is up and running.  Thus, there is no point where the
main thread can return from async_assert and try to send a packet
before tcpdump is operational and ready to receive packets.

Signed-off-by: Mike Micucci <micucci@midokura.com>
Signed-off-by: Duarte Nunes <duarte@midokura.com>
Change-Id: Ie9222e566a3dac06b34cf96444c9a97095b4efbe