Update patch set 2
Patch Set 2:

(1 comment)

WRT the hard-coded classes. I'm not sure how those can be tested, the jersey documentation itself suggests adding their names to the config file, in a way that's equivalent to what the code here is doing:

https://jersey.java.net/apidocs/1.10/jersey/com/sun/jersey/api/container/filter/LoggingFilter.html

If they ever change that they are going to break a lot of people's config files. In fact the ServletContextHandler has methods to add Filters and init params (which would otherwise come from the config file) but it offers no way to set up Container[Request|Response]Filters, of which LoggingFilter is one, so I can't add that programmatically either.

If we are uneasy about this, I can just make sure proper logging is in place everywhere necessary in the ConfMinion and just remove those filters/init-params.

WRT testing in general, you are right. I'll add and auth test using a mock auth service.

Patch-set: 2