Make midonet-api able to run compat api
For now this is just a development tool. The goal is to be able to run
the current midonet-api tests against the compatibility API.  This is
not expected to pass all tests for several reasons, mainly that the
validations and auth made in the compat api are not exactly the same as
in the legacy api.  This is not a problem as the main goal of the compat
API is to allow running MDTS on the new storage stack. However, it is
useful for developers as it allows verifying functionality much faster
than by running MDTS tests.

To run the midonet-api with the new storage stack:

    ./gradlew midonet-api:test -DwithVladimir=true

Note that only a third of the tests do pass right now, this will be
improved in following patches.

This patch also makes the root path of the API configurable.

Change-Id: Ic7e97177be0bbe23fa6c84451014b779f064e74d
Signed-off-by: Galo Navarro <galo@midokura.com>