Make test more robust against env slowness
The test was waiting for an onNext and then asserting the onNext and the
onComplete. Instead, we wait on the onComplete, and then assert both
onNext and onComplete notifications. This should help close a race when
the assert for the onComplete happens before the observable had a chance
to notify.

Change-Id: Ia1e5b3bf79766fd3bc97121fbf9a1d5f39e3f351
Fixes: MNA-383
Signed-off-by: Galo Navarro <galo@midokura.com>