Fixes PoolMember initialization for ZoomConvert
The patch enables ZoomConvert to convert from Protocol Buffer
messages using non-default class constructors. To this end, it
introduces a new @Zoom annotation, which specified the class
constructor to use for setting parameters.

The patch ensures that PoolMember instances, which initialize
private members in the constructor, are converted correctly when
using ZoomConvert.

Fixes: MNA-384

Change-Id: I83f40aded2ce919f32fa284419871bfaf37af1d4
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>