Update patch set 2
Patch Set 2: Code-Review-1

Hi Ivan, thanks for preparing this! As commented off-line, we should  have a way to agree on field names and the values used to represent them outside midolman (maybe having the schema in nsdb)?
The same applies for the values that are internally case classes...

Apart from this, we might also have some test to verify that the field names used by json recorder are the same used by binary recorder and make sure that a change of names does not break anything.

Patch-set: 2
Label: Code-Review=-1