Remove asynchrony from haproxy hm test
The haproxy previously ran all actors in the default akka execution
context. This exposed us to flakiness due to timing differences in each
run. By using TestActorRef, we can make messages to the actors run on
the same thread, and remove the need to use eventually.

However we do still need eventually in some cases as there are periodic
health checks which we are asserting on.

Ref MNA-273

Change-Id: I6306966565fcc91b672de5e2b7120a12dfb946b9
Signed-off-by: Ivan Kelly <ivan@midokura.com>