[WIP] KafkaStorage
This patch implements the MergedMapStorage trait based on Kafka.
Opinion addition and removals are published on the topic that
corresponds to the map.

Kafka message keys contain the map key with the owner that
published this opinion. In conjunction with Kafka's log
compaction feature, we get automatic log cleanup and are
always guaranteed to keep the latest opinion for each
owner and key!

Ref: MNA-367

Change-Id: I6800543fe53dc8596b79d7fabfc1d0a7571c9d13
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>