Kafka Opinion Bus
This patch implements the MergedMapStorage trait based on Kafka.
Opinion addition and removals are published on the topic that
corresponds to the map.

Kafka message keys contain the map key with the owner that
published this opinion. In conjunction with Kafka's log
compaction feature, we get automatic log cleanup and are
always guaranteed to keep the latest opinion for each
owner and key!

Ref: MNA-367

Change-Id: Ic324f0d0ee4ae2b0f4760bdd2fff36acaa24a40e
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>