Fix a flaky test in HostMapperTest
This patch fixes the following flaky test:
Feature: A host should come with its tunnel zones membership
Scenario: The host is in one tunnel zone and then in two

In this test the observer should receive two notifications. We now wait
for the 1st one before waiting for the 2nd one to avoid missing the 1st
notification.

Ref: MNA-381

Change-Id: I6e1377d7a505ed44a251cac4091eea51a1f68c3c
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>