Update patch set 3
Patch Set 3:

Just to clarify: the current http exceptions are coupled to low level data stuff, and even ZK exceptions. This is removed. HTTP exceptions should not need dependencies on the low level storage choices.

Of course, anyone should be use the http exceptions. They are moved to cluster for that reason (and possibly later to nsdb or util)

Patch-set: 3