Fix 1/100 chance of fail on PoolHealthMonitorMapManagerTest
One scenario in PoolHealthMonitorMapManagerTest generates random health
monitors, changes the delay value and expects to be notified of the
update. However, the update is only written if the new health monitor is
different from the old one. There's a 1 in 100 chance of the delay value
already being the value that it is being set to, so this test fails 1 in
100 times. This patch fixes this by making sure the new delay value is
different.

Ref MNA-314

Change-Id: I097875541ed78a3255c5d6d8b610274c30f67be7
Signed-off-by: Ivan Kelly <ivan@midokura.com>