Update routes on RouterInterface create/delete
When a tenant router has a gateway via an external network's subnet,
update the tenant router's default gateway route's next hop gateway IP
address when an interface between the external network and an edge
router is created or deleted on that subnet.

This should complete translation of tenant router uplink.

Signed-off-by: Brandon Berg <bberg@midokura.com>
Change-Id: I10f211260b2f1c86c3440d97a560d2b4862c45af