Make DirectoryCallback rx wrapper use StateAccessExceptions
The places where the rx directory callback wrapper is used only expects
StateAccessExceptions and not KeeperExceptions. In the case that a
KeeperException occurs this will be uncaught and end up killing the
thread. This patch fixes this in two ways. First it wraps any
KeeperException coming from the DirectoryCallback wrapper, in a
StateAccessException. Secondly, it adds a wildcard exception handler, to
port activation so the error doesn't propagate to the uncaught exception
handler for the thread.

Change-Id: Iaee694bfb5749b6877248437705a82404bc4fbc3
Signed-off-by: Ivan Kelly <ivan@midokura.com>