Cluster: support drop-in minions
This patch refactors the midonet-cluster execution service to declare
and manage Minions based on annotations.

For context, MidoNet Cluster services are general purpose services that
run control functions inside MidoNet Cluster nodes. Each Service is
typically provided by a "Cluster Minion", which is just an ordinary
Guice service, with some custom spice.

This patch allows any class extending Minion to declare itself as
provider of a Cluster service by annotating itself as follows:

    @ClusterService(name ="example")
    class ExampleMinion extends Minion

ExampleMinion would be declaring itself as a provider of service
"example".

This change enables vendors to provide custom services by packaging them
in an independent Jar file and dropping them in the midonet-cluster
classpath (/usr/share/midonet-cluster/dep).

Lifecycle
---------

The Cluster Daemon will scan the classpath for all classes annotated
with @ClusterService, and start them asynchronously.  If any of the
minions startup successfully the node will keep running until
termination.  If all of the minions fail to start, then the node will
shut down.

Should two Minions declare themselves as providers of a given
ClusterService, the Daemon will use whatever policy is set in the
@ClusterService annotation:

    @ClusterService(name ="example", priority=Priority.YIELD)
    class ShyExampleMinion extends Minion

    @ClusterService(name ="example", priority=Priority.PREVAIL)
    class BoldExampleMinion extends Minion

In this case, if both BoldExampleMinion and ShyExampleMinion are in the
classpath, only the former would be started.

Given N Minions declaring themselves as implementers of the same
service, the Cluster Node will expect one and only one of them to have a
PREVAIL priority.  If all have YIELD, or more than one have PREVAIL, the
service will not be available in the node.

Change-Id: Iccb953ab19a094ebc005bc4d143690a6300fdadc
Signed-off-by: Galo Navarro <galo@midokura.com>