Delete bug in Replicated Maps
This patch addresses a bug in the replicated maps where an entry would
never be deleted if the 1st attempt fails in method
removeIfOwnerAndValue. Indeed, in the previous implementation the entry
version was removed from map ownedVersions before performing the
corresponding deletion in zookeeper. If the latter operation failed,
then the version would not be owned by the client and the map entry
would therefore never be added to cleanupPaths in method
populateNewMap(). As a consequence, the entry would remain in Zookeeper
forever.

Ref: MN-930

Change-Id: Ia5f334fbf49f67cdd90c86727d427394cbf87c53
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>