Agent: Fetch flow state upon port activation
This patches fixes a regression which prevented us from fetching flow
state from Cassandra.

Since fetching the flow state associated with a port is now performed
asynchronously to activating that port, we retry checking the return
flow in the test_distributed_l4_port_binding so that we allow the
agent to connect to and fetch the relevant state form Cassandra. When
this state is fetched, we invalidate the flows that are tagged with
that state

Change-Id: I6854d4e5fa591e8b2883e726c8d7f2f91a05fc75
Signed-off-by: Duarte Nunes <duarte@midokura.com>