Update patch set 1
Patch Set 1:

If the call to ZK fails, we will delete the node only if a fresher version is put in the map. So you're right, in the case this does not happen, the node remains. I'll resubmit a new version of this patch that fixes this.

Patch-set: 1
Label: Code-Review=0