MDTS: Fix zookeeper failure handling
Currently, zookeeper failures are simulated by setting the interface
for the corresponding namespace down. MDTS wait s a fixed amount of
time before continuing the test. This may have several issues like
starting a new test with on zk host down (no proper isolation) or
midonet_api failing because the connection to that zk host was
not properly restored.

This fix includes a new type of failure (Zookeeper failure) which
makes sure of:
- the failing zk host is indeed down before starting the test
- the zk host is indeed brought up correctly after the test and
  before continuing with subsequent tests.

Some code cleanup and minor reorganization.

Fixes: MN-1648
Change-Id: I10859b417b0203144a460eda4a948ff6ee863f89