VxGW: remove macs when vxlan port removed
This patch will ensure that whenever a VxLAN port is removed from a
bridge (this happens when the last VTEP binding is removed) all the
MAC-Port entries associated to the VxLAN port (that is, all MACs in the
VTEP) are removed from the MacPort map, as well as the MAC-IP entries
from the ArpTable.

This adds minor related fixes to the tests, and also a fix handling
mac-port replacements, as the replicated maps were leaving entries
behind even when doing a replacement.

It also solves a race on VxLAN port removal, correctly cleaning up state
and notifying the VTEP controller.

Ref.: MN-3556
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I7e8c3f189a5d48684c9f1b52c7a775532d6af979