Delete bug in Replicated Maps
This patch addresses a bug in the replicated maps where an entry would
never be deleted if the 1st attempt failed in method
removeIfOwnerAndValue. This is simply because if the deletion of
the ZK node fails, no retry is attempted.

To solve this issue, if the operation fails due to an error or
a timeout, the operation is re-issued. In addition the version
to be deleted is only removed from the owned versions when
the delete operation is successful.

Ref: MN-930

Change-Id: I7fcff588048b2eb8b41be0b641931b89c1634784
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>