MDTS: wait port inactive after unbind and clear bgp
At the time of unbinding, we do not wait for the port to be inactive so
if we go through to the next test and try to bind the interface before
the agent actually unbinds it, we may hit the error "Interface is
already bound to a port". The fix is to wait for the port to be
inactive.

This also fixes BGP tests that may fail because the default route is
not advertised on time (60 seconds increased to 120). That may cause
other tests to fail trying to bind to an interface stil bound to a port.

Change-Id: I8a516deb49403502cb64a66fd1b0821364b9b343