API: decouple neutron APIs from old storage stack
This patch removes all the backend exceptions from the Neutron plugin
API replacing them with HTTP exceptions.  The reasoning behind this
change is to enable a new implementation of the plugin api that supports
the new storage layer (which doesn't share the same exceptions as the
old one).

Exceptions are now converted into HttpExceptions, this might not be the
best option if the users of the plugin API are not HTTP frontends, but
since this is so far very unlikely, so it simplifies to assume that this
plugin API will be used by REST APIs.

Ref.: MNA-419
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I94513c82e281a06fb6c463d361691221da8410e3