Allow priorities to be assigned to selectable channels
We observed a bug where flow state messages were arriving at a agent
before a ping reply that required it even though kernel traced showed
that the flow state message was arriving in the network state and being
sent up to the agent before the ping reply. It turns out that the order
in which the agent reads from it's selectable channels is pretty much
random, so there's there was no way to guarantee flow state is handled
first.

This patch adds a concept of priorities to channels. For now there are
only two priorities, HIGH and NORMAL. HIGH is only given to tunnel
channels which are the channels which could be carrying flow state.

I've also removed retry logic from test_bgp.py which was only there
to (unsuccessfully) work around the issue.

Ref MNA-337

Change-Id: I945246d2d0e2c1a1ca10e2edc790ad38afc36cef
Signed-off-by: Ivan Kelly <ivan@midokura.com>