ZOOM: Enhancing support for ephemeral data
This patch introduces a `StorageWithState` trait with the objective
of enhancing the support for ephemeral state data. Currently, ZOOM
only supports ephemeral in the form of ownership metadata, which
can be added to existing objects. However, this model has several
drawbacks:
- Concurrent modifications to the topology and state data may result
in some operations to fail. Handling this failures increases the
complexity around the storage write operations.
- In the current design, ZOOM erases the ownership when an object is
modified using the referential integrity bindings.
- The ownership model does not accommodate well certain types of
state data, such as BGP status information or whether a route is
learned.

Ref: MNA-397

Change-Id: I5e86a62884683db0578770685c7bccd8e42be3f7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>