Add ENABLE_API/ENABLE_CLUSTER vars in devmido
To be able to develop against both the API and the Cluster versions of
midonet, provide options to start both services.

This patch deprecates USE_CLUSTER variable but because all services
start by default, anything that used to depend on USE_CLUSTER to start
the cluster would still work.

The new environment variables introduced are:

 * ENABLE_API: enables the API service
 * ENABLE_CLUSTER: enables the cluster services

They both default to True, and so does ENABLE_TASKS_IMPORTER.

Change-Id: I66faa1e15637e31159c0cc5c42328c32785f45fc
Implements: MNA-427