ZOOM: Introduce NodeObservable
This class improves the performance of the existing
ObservableNodeCache by providing the following:
- A watcher is installed on the getData rather than exists, such
that the ZK server does not install a watcher on non-existing
nodes. However, the option of installing an exists watcher is
maintained through an observable switch, which gives the capability
to issue on-next notifications when the path is deleted, rather
than complete the observable.
- Improves the error handling such that we process error for
both the watcher callback and the background callback of the
asynchronous operation.
- Provides a cleanup mechanism, which clears watcher references
and connection state listeners upon the last unsubscribe. This also
ensures that the observable enters a terminal state, such that
subscribers that still hold references to it may gracefully recover
to a new observable.

This implementation is intended to replace the ObservableNodeCache
in the ZOOM client.

Ref: MNA-129
Ref: MNA-397

Change-Id: I60572194c7b715f3ab96cd614e1100e3ba5308c7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>