MMM: integrate mn-conf (allow global config)
This patch allows the execution of an script just after
setting up the zookeeper cluster. By default, it will execute
the (empty) file in mmm/scripts/mn-conf/mn-conf.sh

The script to execute can be defined through the
MN_CONF_SCRIPT env var in case an out-of-tree script
should be executed.

Fixes: MT-420
Change-Id: Iecf00b7508592ded396b99c39e34ef73b444d94d