devmido: force jettyRun to use web.xml
build.gradle of midonet-api specified web.xml.dev as the configuration
file when running jettyRun. In devmido, it still thinks that web.xml is
the configuration file. It is undesirable to modify web.xml.dev since
it is meant to be a sample file and source controlled. To get around
this, specify explicitly web.xml in jettyRun command.

Change-Id: Ib1ce5077febc1bae9d67b1eaaf95c2a9eece35d6
Fixes: MNA-469