Agent: Fixes to dealing with an existing DP port
When creating a DP port, we handle the case where that port already
exists. One reason it may exist can be because we timed out a previous
creation attempt, due to, for example, a long GC; netlink timeouts
don't really exist, so the operation still completed.

This patch adds an exception - address already in use - that tells us the port
exists in case of a Vxlan tunnel port.

It also changes the recovery to delete and re-create the port, since
simply setting the upcall pid fails with operation not supported for
Vxlan tunnel ports.

Change-Id: If834f0bc4df235a8b726c14f41b3670621967d4a
Signed-off-by: Duarte Nunes <duarte@midokura.com>