Agent: Packets using 802.1Q are not VLAN tagged
Agent: Route VLAN 0 tagged packets

This patch ensures we allow packets tagged with VLAN 0 through the
router, because this tag only specifies a priority. Packets use VLAN 0 to
specify the Class of Service, for which there are bits in the Ethernet
header. VLAN aware or unaware bridges treat such packets as if
untagged, and so now does our virtual router.

Ref MNA-222

Change-Id: I725132817da81f0b1ce437c75bce4355de723d72
Signed-off-by: Duarte Nunes <duarte@midokura.com>