Compat API: perform validations and fix Route DTO
This patch enables the validation checks for CREATE and UPDATE
operations in the Compat API, fixing an issue spotted in the Route DTO
that didn't have a ZoomField annotation in the weight.

Since we add a validator that must be injected in every resource along
with backend and uriInfo, we also pack these into a container
(ResourceContext).  This will make the code more maintainable, as new
deps. that require injection into any resources can be added to this
single place, and used wherever necessary, without requiring to add a
new parameter to every resource class.

Change-Id: I90eb5230594181b8968bb1334fe91a6c05e858ad
Signed-off-by: Galo Navarro <galo@midokura.com>