Fix Test/AwaitableObserver flakiness
Several tests rely on an observer that mixes in TestObserver (from the
rx package) and AwaitableObserver. The order in which onNext is called
on an observer depends on the order in which TestObserver and
AwaitableObserver appear in the observer class used.  For class
DeviceObserver, TestObserver previously appeared before
AwaitableObserver, which meant that onNext in AwaitableObserver was
called before the same method in TestObserver. As a result, a test
could return from an awaitOnNext call before the emitted item was
placed in the onNextEvents list of TestObserver.

This patch solves this problem by not relying on TestObserver anymore
and placing the functionality we were using of that class in
AwaitableObserver.

Ref: MNA-347

Change-Id: Ie842a4701d4459488ecf7356323fbc1f9fb2544a
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>