Compat API: Fix DHCP and VIP unit tests
This makes some midonet-api tests pass with the compat api, by tweaking
the response code from 200 to 204 for consistency, also handling cases
where a list of subresources was serialized as an empty array, but
expected as a null (in ZOOM we do store an empty list, so it won't
return null in the DTO even if no elements were set).

Also fixes a bunch of VIP tests, by actually solving a bug in the API
translations.

Pending to this patch is to refactor some of the assertions in the L4LB
tests to accept 400 when not found, rather than 404 with customized
error messages.

Change-Id: I0d07bfc73e000cb2e5f9a731b7b889b093b08a7e
Signed-off-by: Galo Navarro <galo@midokura.com>