Update patch set 1
Patch Set 1:

(14 comments)

Hi Brandon, Tomohiko, Thanks for the thorough review. I have addressed your comments. Also, in a discussion Brandon pointed out that a lot of the code re-arranging is not necessary if I don't try messing with the SG in the SGR translator. I have made this change as well and as a result the patch is much smaller.

Patch-set: 1
Label: Code-Review=0