ReplicatedMap fixes
This patch avoids a double notification when removing an entry in the
map and improves logging in the delete callback.

Previously, we would notify the watcher of the entry deletion in the
delete callback. However, this could cause a double notification.
Indeed, the deletion may have been notified previously in the
directory watcher when collecting notifications.

Change-Id: I58e6577413f2e43ca110459d27edbf89bc7bdece
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>