Remove get from NeutronZoomPlugin create/update
NeutronZoomPlugin.create() and .update() were performing a get() to
retrieve the created/updated object from storage. This is unnecessary
(since C3PO stores the objects as passed, and thus the object returned
from get() will always be the same as the one just passed to
create()/update()), and it was causing an exception for
addRouterInterface(), because we don't store the Neutron
RouterInterface objects in Zoom. This caused the get() to fail and
throw an exception.

Change-Id: I13de70b818bb5cd8893421a2ac2abf13ef39eea9
Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MNA-503