Remove get from NeutronZoomPlugin create/update
NeutronZoomPlugin.create() and .update() were performing a get() to retrieve the created/updated object from storage. This is unnecessary (since C3PO stores the objects as passed, and thus the object returned from get() will always be the same as the one just passed to create()/update()), and was causing an exception for addRouterInterface(), because we don't store the Neutron RouterInterface objects in Zoom. This caused the get() to fail and throw an exception.

Change-Id: I147614daa5e5bb6af4d7fbfc1f6d3f85d86ef144
Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MNA-503