State: State storage support for learned routes
The patch introduces utility methods providing support for saving
learned routes to the state storage. These are implemented as a
wrapper class around StateStorage, providing the necessary seriali-
zation support for routes, and leveraging the state storage
key-value mechanism.

The patch also adds a routing table benchmark and a set of results
estimating the routing table performance for a same host scenario.

In addition, it renames the agent StateStorage to LegacyStorage and
moves it to the LegacyClusterModule.

Ref: MNA-530

Change-Id: I5446964f10f55c0b6e68160723e5e43dc15c0f21
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>