Define UPLINK network type
Previously, 'LOCAL' network type is used to determine whether a Neutron
network is an uplink network or not. This causes a problem in Neutron
because all admin-owned networks must have a type, and it always
defaults to 'LOCAL', making it impossible to differentiate uplink and
normal networks.  To get around it, just define a new type called
'UPLINK' to replace 'LOCAL'.

Change-Id: Idcb8279cbcc61f45d0626a2a73c4d011766b2c71
Implements: MNA-526