Expose reason for assert failure
This patch replaces an assertion from (essentially) x = y should be true to a x
should be y, so that upon failure we can spot the value being thrown.

This is trying to identify the underlying cause behind a flake that is
likely to be caused by a race condition, but never reproduces locally.

Ref.: MNA-510
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Ie957382d91b3a6eb53695634296aed6494b71d36