Update patch set 8
Patch Set 8:

Thanks Nicolas! I will address your comments in a subsequent patch, in order to unblock them.

Related to your suggestion to merge tests: The ZK has one feature that cannot be accurately tested with a mock: the session ID and the ability to distinguish between clients. These checks are missing for the in-memory version. In theory we could consolidate the common code in a single class, but in the end I tough it might be too messy. It's worth giving a try though.

Patch-set: 8