Agent: Don't clear flow match on suspension
This patch ensures we don't clear the modified flow match when we
suspend a packet, because if we evict the packet from the waiting room
and drop it, we'll install a drop flow that needs the changes to the
flow match in order to correctly calculate the flow mask. Otherwise,
this mask would essentially be a very wide megaflow that would cause
extra drops during the 5 seconds the flow is installed.

Ref MNA-546

Change-Id: I6a3e158454b249010f02bffcef40c628257aaee4